Make things closer to paper
