+++ benchmark_results.txt
@@ -1,54 +1,65 @@
-01_basic.rs: 52028438
-02_list_zero.rs: 85853737
-03_unnest.rs: 60505988
-04_chunks.rs: 168033346
-05_is_xid_start.rs: 174421652
-06_cursor.rs: 96822026
-07_fmt.rs: 122985824
-08_parse_parens.rs: 69940626
-09_punct.rs: 308323927
-10_members.rs: 68700251
-11_display.rs: 83587203
-12_jasper.rs: 86364549
-13_heap_visitor.rs: 62454191
-14_serde_err.rs: 3462187912
-15_bernoulli.rs: 104690104
-16_lookahead.rs: 64083431
-17_entry.rs: 84072845
-18_slice.rs: 76896282
-19_lifetime_projection.rs: 59955020
-20_eager.rs: 58015853
-21_jasper2.rs: 69523839
-22_itoa_format.rs: 77597466
-23_lazy_poll.rs: 78995035
-24_style_render_reset.rs: 162262409
-25_parse_str.rs: 65765888
-26_ref_in_struct.rs: 64201579
-27_aurel.rs: 58627398
-28_ranged_slice.rs: 80487100
-29_prusti_issue_25.rs: 67626258
-30_prusti_issue_738-5.rs: 63542806
-31_prusti_timeout.rs: 106783199
-32_deflate64_inputbuffer_advance.rs: 86405375
-33_hashtable_shrink_to_fit.rs: 70102835
-34_pair_into_value.rs: 45165834
-35_miniz_oxide_new_output_buffer.rs: 166735732
-36_gimli_endian_slice.rs: 188537238
-37_demo_loop.rs: 114773638
-38_regex_autonoma_iter_pattern_match_ids.rs: 98097128
-39_cfi.rs: 107803323
-40_object_archive.rs: 207285530
-41_itertools_k_smallest.rs: 156442180
-42_dfa_dense.rs: 2037997221
-43_der_length.rs: 123287424
-44_rand_weighted_index.rs: 3543035131
-45_closure.rs: 60129282
-46_demo_cond_mut_borrow.rs: 63269976
-47_demo_choose.rs: 57590128
-48_flowistry_basic.rs: 63449909
-49_flowistry_recurse.rs: 46184667
-50_flowistry_pointer_deep.rs: 62089829
-51_flowistry_recurse_not_all_args.rs: 64806547
-52_flowistry_recurse_parent_privacy.rs: 64456075
-53_flowistry_time_calculation.rs: 75482269
-54_flowistry_struct_mut_ptr.rs: 63412486
+01_basic.rs: 52112473
+02_list_zero.rs: 87472937
+03_unnest.rs: 61316231
+04_chunks.rs: 239663879
+05_is_xid_start.rs: 174222132
+06_cursor.rs: 99560963
+07_fmt.rs: 138767342
+08_parse_parens.rs: 70424921
+09_punct.rs: 277169380
+10_members.rs: 68449887
+11_display.rs: 84230773
+12_jasper.rs: 90182101
+13_heap_visitor.rs: 62255094
+14_serde_err.rs: 1443513682
+15_bernoulli.rs: 105749762
+16_lookahead.rs: 64041096
+17_entry.rs: 85993613
+18_slice.rs: 90235643
+19_lifetime_projection.rs: 60626056
+20_eager.rs: 58086317
+21_jasper2.rs: 73969353
+22_itoa_format.rs: 77187941
+23_lazy_poll.rs: 78772950
+24_style_render_reset.rs: 156767346
+25_parse_str.rs: 65618424
+26_ref_in_struct.rs: 66534201
+27_aurel.rs: 58769991
+28_ranged_slice.rs: 87176778
+29_prusti_issue_25.rs: 69710662
+30_prusti_issue_738-5.rs: 65294475
+31_prusti_timeout.rs: 106792360
+32_deflate64_inputbuffer_advance.rs: 90816797
+33_hashtable_shrink_to_fit.rs: 69663850
+34_pair_into_value.rs: 45084754
+35_miniz_oxide_new_output_buffer.rs: 184662373
+36_gimli_endian_slice.rs: 203562088
+37_demo_loop.rs: 116691222
+38_regex_autonoma_iter_pattern_match_ids.rs: 111065521
+39_cfi.rs: 112508164
+40_object_archive.rs: 246828462
+41_itertools_k_smallest.rs: 141438965
+42_dfa_dense.rs: 2020726965
+43_der_length.rs: 126033255
+44_rand_weighted_index.rs: 1633537216
+45_closure.rs: 60334125
+46_demo_cond_mut_borrow.rs: 63837039
+47_demo_choose.rs: 60074638
+48_flowistry_basic.rs: 64435600
+49_flowistry_recurse.rs: 46399314
+50_flowistry_pointer_deep.rs: 64706057
+51_flowistry_recurse_not_all_args.rs: 65247317
+52_flowistry_recurse_parent_privacy.rs: 64586429
+53_flowistry_time_calculation.rs: 86384118
+54_flowistry_struct_mut_ptr.rs: 64575793
+55_prusti_ownership2.rs: 78372193
+56_deep2.rs: 60963247
+57_slice_ptr_elem_write.rs: 61586124
+58_aurel_pledge.rs: 64579332
+59_struct_ptrs_deep.rs: 97067407
+60_pointer_reborrow_nested.rs: 62452043
+61_hashset_union.rs: 123759525
+62_tuple_ptr_write_field.rs: 63963905
+63_flowistry_function_lifetime_alias_mut.rs: 59372415
+64_flowistry_enum_write_branch_read_whole.rs: 70499566
+65_flowistry_recurse_project_dst.rs: 46497025
