+++ src/borrow_pcg/edge/borrow.rs
@@ -10,10 +10,10 @@ use crate::{
     },
 };
 
-use crate::borrows::borrow_pcg_edge::{BlockedNode, LocalNode};
-use crate::borrows::edge_data::EdgeData;
-use crate::borrows::has_pcs_elem::HasPcsElems;
-use crate::borrows::region_projection::RegionProjection;
+use crate::borrow_pcg::borrow_pcg_edge::{BlockedNode, LocalNode};
+use crate::borrow_pcg::edge_data::EdgeData;
+use crate::borrow_pcg::has_pcs_elem::HasPcsElems;
+use crate::borrow_pcg::region_projection::RegionProjection;
 use crate::utils::display::DisplayWithRepacker;
 use crate::utils::json::ToJsonWithRepacker;
 use crate::utils::place::maybe_old::MaybeOldPlace;
