+++ src/utils/eval_stmt_data.rs
@@ -1,6 +1,6 @@
 use serde_json::json;
 use crate::utils::json::ToJsonWithRepacker;
-use crate::borrows::engine::BorrowsStates;
+use crate::borrow_pcg::engine::BorrowsStates;
 use crate::combined_pcs::EvalStmtPhase;
 use crate::utils::PlaceRepacker;
 use crate::utils::validity::HasValidityCheck;
