+++ src/visualization/graph_constructor.rs
@@ -1,8 +1,8 @@
 use crate::{
-    borrows::{
+    borrow_pcg::{
         borrow_pcg_edge::{BorrowPCGEdgeLike, LocalNode},
-        borrows_graph::BorrowsGraph,
-        borrows_state::BorrowsState,
+        graph::BorrowsGraph,
+        state::BorrowsState,
         coupling_graph_constructor::CGNode,
         region_projection::{MaybeRemoteRegionProjectionBase, RegionProjection},
         unblock_graph::UnblockGraph,
@@ -17,8 +17,8 @@ use crate::{
 };
 
 use super::{dot_graph::DotSubgraph, Graph, GraphEdge, GraphNode, NodeId, NodeType};
-use crate::borrows::edge::abstraction::AbstractionType;
-use crate::borrows::edge::kind::BorrowPCGEdgeKind;
+use crate::borrow_pcg::edge::abstraction::AbstractionType;
+use crate::borrow_pcg::edge::kind::BorrowPCGEdgeKind;
 use crate::rustc_interface::middle::ty::TyCtxt;
 use crate::utils::place::maybe_old::MaybeOldPlace;
 use crate::utils::place::maybe_remote::MaybeRemotePlace;
